refactor(l5): move embedding-repair from src/ into server/embedding (issue 864)
Old -> new:

  src/embedding-repair.ts -> server/embedding/embedding-repair.ts
                             server/embedding/embedding-repair-selection.ts
                             server/embedding/embedding-repair-model.ts

src/embedding-repair.ts stays as an L5 re-export shim so every existing
caller keeps working; it is retired with src/ at L6.

The test file src/embedding-repair.test.ts stays at its src/ path,
untouched and unstaged, per the M1 exception: transforming a 963-line
test with 54 findings does not fit this lane, and it exercises the moved
code through the shim. src/embedding-repair.pg.test.ts,
src/embedding-repair-formula.pg.test.ts, and
src/embedding-repair-test-helpers.ts stay in src/ for the same reason.

No behavior change.
